+++ crates/gpui/src/platform/windows/platform.rs
@@ -3,7 +3,7 @@ use std::{
     mem::ManuallyDrop,
     path::{Path, PathBuf},
     rc::Rc,
-    sync::{Arc, LazyLock},
+    sync::Arc,
 };
 
 use ::util::ResultExt;
@@ -16,25 +16,12 @@ use windows::{
     core::*,
     Win32::{
         Foundation::*,
-        Globalization::u_memcpy,
         Graphics::{
             Gdi::*,
             Imaging::{CLSID_WICImagingFactory, IWICImagingFactory},
         },
         Security::Credentials::*,
-        System::{
-            Com::*,
-            DataExchange::{
-                CloseClipboard, EmptyClipboard, GetClipboardData, GetPriorityClipboardFormat,
-                IsClipboardFormatAvailable, OpenClipboard, RegisterClipboardFormatW,
-                SetClipboardData,
-            },
-            LibraryLoader::*,
-            Memory::{GlobalAlloc, GlobalLock, GlobalSize, GlobalUnlock, GMEM_MOVEABLE},
-            Ole::*,
-            SystemInformation::*,
-            Threading::*,
-        },
+        System::{Com::*, LibraryLoader::*, Ole::*, SystemInformation::*, Threading::*},
         UI::{Input::KeyboardAndMouse::*, Shell::*, WindowsAndMessaging::*},
     },
     UI::ViewManagement::UISettings,
@@ -50,8 +37,7 @@ pub(crate) struct WindowsPlatform {
     background_executor: BackgroundExecutor,
     foreground_executor: ForegroundExecutor,
     text_system: Arc<DirectWriteTextSystem>,
-    clipboard_hash_format: u32,
-    clipboard_metadata_format: u32,
+    clipboard_formats: ClipboardFormatStore,
     windows_version: WindowsVersion,
     bitmap_factory: ManuallyDrop<IWICImagingFactory>,
     validation_number: usize,
@@ -104,9 +90,8 @@ impl WindowsPlatform {
         let icon = load_icon().unwrap_or_default();
         let state = RefCell::new(WindowsPlatformState::new());
         let raw_window_handles = RwLock::new(SmallVec::new());
-        let clipboard_hash_format = register_clipboard_format(CLIPBOARD_HASH_FORMAT).unwrap();
-        let clipboard_metadata_format =
-            register_clipboard_format(CLIPBOARD_METADATA_FORMAT).unwrap();
+        let clipboard_formats =
+            ClipboardFormatStore::new().expect("Unable to register clipboard formats");
         let windows_version = WindowsVersion::new().expect("Error retrieve windows version");
         let validation_number = rand::random::<usize>();
 
@@ -117,8 +102,7 @@ impl WindowsPlatform {
             background_executor,
             foreground_executor,
             text_system,
-            clipboard_hash_format,
-            clipboard_metadata_format,
+            clipboard_formats,
             windows_version,
             bitmap_factory,
             validation_number,
@@ -466,15 +450,11 @@ impl Platform for WindowsPlatform {
     }
 
     fn write_to_clipboard(&self, item: ClipboardItem) {
-        write_to_clipboard(
-            item,
-            self.clipboard_hash_format,
-            self.clipboard_metadata_format,
-        );
+        write_to_clipboard(item, &self.clipboard_formats);
     }
 
     fn read_from_clipboard(&self) -> Option<ClipboardItem> {
-        read_from_clipboard(self.clipboard_hash_format, self.clipboard_metadata_format)
+        read_from_clipboard(&self.clipboard_formats)
     }
 
     fn write_credentials(&self, url: &str, username: &str, password: &[u8]) -> Task<Result<()>> {
@@ -682,180 +662,6 @@ fn should_auto_hide_scrollbars() -> Result<bool> {
     let ui_settings = UISettings::new()?;
     Ok(ui_settings.AutoHideScrollBars()?)
 }
-
-fn register_clipboard_format(format: PCWSTR) -> Result<u32> {
-    let ret = unsafe { RegisterClipboardFormatW(format) };
-    if ret == 0 {
-        Err(anyhow::anyhow!(
-            "Error when registering clipboard format: {}",
-            std::io::Error::last_os_error()
-        ))
-    } else {
-        Ok(ret)
-    }
-}
-
-fn write_to_clipboard(item: ClipboardItem, hash_format: u32, metadata_format: u32) {
-    write_to_clipboard_inner(item, hash_format, metadata_format).log_err();
-    unsafe { CloseClipboard().log_err() };
-}
-
-fn write_to_clipboard_inner(
-    item: ClipboardItem,
-    hash_format: u32,
-    metadata_format: u32,
-) -> Result<()> {
-    unsafe {
-        OpenClipboard(None)?;
-        EmptyClipboard()?;
-    }
-    for entry in item.entries() {
-        match entry {
-            ClipboardEntry::String(string) => {
-                write_string_to_clipboard(string, hash_format, metadata_format)?;
-            }
-            ClipboardEntry::Image(image) => {
-                write_image_to_clipboard(image)?;
-            }
-        }
-    }
-    Ok(())
-}
-
-fn write_string_to_clipboard(
-    item: &ClipboardString,
-    hash_format: u32,
-    metadata_format: u32,
-) -> Result<()> {
-    let encode_wide = item.text.encode_utf16().chain(Some(0)).collect_vec();
-    set_data_to_clipboard(&encode_wide, CF_UNICODETEXT.0 as u32)?;
-
-    if let Some(metadata) = item.metadata.as_ref() {
-        let hash_result = {
-            let hash = ClipboardString::text_hash(&item.text);
-            hash.to_ne_bytes()
-        };
-        let encode_wide =
-            unsafe { std::slice::from_raw_parts(hash_result.as_ptr().cast::<u16>(), 4) };
-        set_data_to_clipboard(encode_wide, hash_format)?;
-
-        let metadata_wide = metadata.encode_utf16().chain(Some(0)).collect_vec();
-        set_data_to_clipboard(&metadata_wide, metadata_format)?;
-    }
-    Ok(())
-}
-
-fn set_data_to_clipboard(data: &[u16], format: u32) -> Result<()> {
-    unsafe {
-        let global = GlobalAlloc(GMEM_MOVEABLE, data.len() * 2)?;
-        let handle = GlobalLock(global);
-        u_memcpy(handle as _, data.as_ptr(), data.len() as _);
-        let _ = GlobalUnlock(global);
-        SetClipboardData(format, HANDLE(global.0))?;
-    }
-    Ok(())
-}
-
-fn write_image_to_clipboard(item: &Image) -> Result<()> {
-    unsafe {
-        let data = item.bytes();
-        let global = GlobalAlloc(GMEM_MOVEABLE, data.len())?;
-        let handle = GlobalLock(global);
-        std::ptr::copy_nonoverlapping(data.as_ptr(), handle as _, data.len());
-        let _ = GlobalUnlock(global);
-        SetClipboardData(CF_DIB.0 as u32, HANDLE(global.0))?;
-    }
-    Ok(())
-}
-
-fn read_from_clipboard(hash_format: u32, metadata_format: u32) -> Option<ClipboardItem> {
-    let result = read_from_clipboard_inner(hash_format, metadata_format).log_err();
-    unsafe { CloseClipboard().log_err() };
-    result
-}
-
-fn read_from_clipboard_inner(hash_format: u32, metadata_format: u32) -> Result<ClipboardItem> {
-    unsafe {
-        OpenClipboard(None)?;
-        let mut entries = Vec::new();
-        if let Some(string) = read_string_from_clipboard(hash_format, metadata_format).log_err() {
-            entries.push(string);
-        }
-        if let Some(image) = read_image_from_clipboard().log_err() {
-            entries.push(image);
-        }
-        if entries.is_empty() {
-            anyhow::bail!("No content in clipboard");
-        } else {
-            Ok(ClipboardItem { entries })
-        }
-    }
-}
-
-fn read_string_from_clipboard(hash_format: u32, metadata_format: u32) -> Result<ClipboardEntry> {
-    let text = unsafe {
-        let handle = GetClipboardData(CF_UNICODETEXT.0 as u32)?;
-        let text = PCWSTR(handle.0 as *const u16);
-        String::from_utf16_lossy(text.as_wide())
-    };
-    let Some(hash) = read_hash_from_clipboard(hash_format) else {
-        return Ok(ClipboardEntry::String(ClipboardString::new(text)));
-    };
-    let Some(metadata) = read_metadata_from_clipboard(metadata_format) else {
-        return Ok(ClipboardEntry::String(ClipboardString::new(text)));
-    };
-    if hash == ClipboardString::text_hash(&text) {
-        Ok(ClipboardEntry::String(ClipboardString {
-            text,
-            metadata: Some(metadata),
-        }))
-    } else {
-        Ok(ClipboardEntry::String(ClipboardString::new(text)))
-    }
-}
-
-fn read_hash_from_clipboard(hash_format: u32) -> Option<u64> {
-    unsafe {
-        let handle = GetClipboardData(hash_format).log_err()?;
-        let raw_ptr = handle.0 as *const u16;
-        let hash_bytes: [u8; 8] = std::slice::from_raw_parts(raw_ptr.cast::<u8>(), 8)
-            .to_vec()
-            .try_into()
-            .log_err()?;
-        Some(u64::from_ne_bytes(hash_bytes))
-    }
-}
-
-fn read_metadata_from_clipboard(metadata_format: u32) -> Option<String> {
-    unsafe {
-        let handle = GetClipboardData(metadata_format).log_err()?;
-        let text = PCWSTR(handle.0 as *const u16);
-        Some(String::from_utf16_lossy(text.as_wide()))
-    }
-}
-
-static PNG_FORMAT: LazyLock<u32> =
-    LazyLock::new(|| unsafe { RegisterClipboardFormatW(windows::core::w!("PNG")) });
-
-fn read_image_from_clipboard() -> Result<ClipboardEntry> {
-    unsafe {
-        let global = HGLOBAL(GetClipboardData(*PNG_FORMAT)?.0);
-        let image_ptr = GlobalLock(global);
-        let iamge_size = GlobalSize(global);
-        let bytes = std::slice::from_raw_parts(image_ptr as *mut u8 as _, iamge_size).to_vec();
-        let _ = GlobalUnlock(global);
-        let id = hash(&bytes);
-        Ok(ClipboardEntry::Image(Image {
-            format: ImageFormat::Png,
-            bytes,
-            id,
-        }))
-    }
-}
-
-// clipboard
-pub const CLIPBOARD_HASH_FORMAT: PCWSTR = windows::core::w!("zed-text-hash");
-pub const CLIPBOARD_METADATA_FORMAT: PCWSTR = windows::core::w!("zed-metadata");
 
 #[cfg(test)]
 mod tests {
